refactor state related actions into states package from reconciler
